Cleaning code and misc
Separating Config.hpp code into several files: SimTaDyn.hpp and
PathManager.hpp.
Improving SimTaDyn config: for example log file was created near the
exec, now in /tmp).
Moving main.cpp code into SimTaDynContext class.
Logger remove dependencies with File and Config because could not add
LOGx() inside File.cpp code.
Remove gtksourceviewmm compilation warnings.
Remove useless includes.
Add class for defining strong type (like using foo = int which is just
an alias)

+++ src/common/patterns/ClassCounter.tpp
@@ -25,7 +25,7 @@
 #  include <cstddef>
 #  include <atomic>
 
-typedef uint32_t Key;
+typedef size_t Key;
 
 // *************************************************************************************************
 //! \brief This class allows to count the number of created
